implements HTTP method DELETE, handles HttpStatus.NOT_FOUND

+++ src/main/java/Java_practise/H2/DB/learning/service/StudentServiceImpl.java
@@ -3,7 +3,9 @@ package Java_practise.H2.DB.learning.service;
 import Java_practise.H2.DB.learning.entity.Student;
 import Java_practise.H2.DB.learning.repository.StudentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -39,7 +41,15 @@ public class StudentServiceImpl implements StudentService {
 
     @Override
     public void deleteStudent(Long studentId){
-         studentRepository.deleteById(studentId);
+        if (studentRepository.existsById(studentId)) {
+            studentRepository.deleteById(studentId);
+        }
+        else{
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND,
+                    "Student with id " + studentId + " not found");
+                    }
+
     }
